Rework GasTracker types + use `LedgerExplorerId`
