+++ libs/ledgerjs/packages/cryptoassets/src/currencies.ts
@@ -3600,10 +3600,6 @@ export const cryptocurrenciesById: Record<CryptoCurrencyId, CryptoCurrency> = {
   },
   // ethereum nanoapp currencies
   // Light Integrations are at the end of the list until we figure out a way to fix the ticker/managerApp collisions
-  /**
-   * FIXME: cleanup libs/coin-evm/src/api/gasTracker/ledger.ts currencyIdGasTrackerMap
-   * when ethereum_as_evm_test_only is removed from the currencies list
-   */
   ethereum_as_evm_test_only: {
     type: "CryptoCurrency",
     id: "ethereum_as_evm_test_only",
@@ -3632,10 +3628,6 @@ export const cryptocurrenciesById: Record<CryptoCurrencyId, CryptoCurrency> = {
     ],
     explorerId: "eth",
   },
-  /**
-   * FIXME: cleanup libs/coin-evm/src/api/gasTracker/ledger.ts currencyIdGasTrackerMap
-   * when polygon_as_evm_test_only is removed from the currencies list
-   */
   polygon_as_evm_test_only: {
     type: "CryptoCurrency",
     id: "polygon_as_evm_test_only",
